runners: jlink: fix version parsing
JLink V7.0a does not conform to PEP 440 version conventions; the 'a'
part is used by PEP 440 compliant versions for alphas. It gets parsed
to a legacy type by the packaging library, which always is treated as a
lower value when compared with a conforming version string.

To fix, parse the version manually with something J-Link specific.

Reported-by: Jake Mercer <jake.mercer@civica.co.uk>
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>